+++ examples/sending-mime-email-with-freemarker-example/src/main/java/com/test/TestService.java
@@ -18,6 +18,7 @@ import org.springframework.stereotype.Service;
 import javax.mail.internet.InternetAddress;
 import java.io.File;
 import java.io.UnsupportedEncodingException;
+import java.net.URISyntaxException;
 import java.nio.charset.Charset;
 import java.util.Map;
 
@@ -29,7 +30,7 @@ public class TestService {
     @Autowired
     private EmailService emailService;
 
-    public void sendMimeEmailWithFreemarker() throws UnsupportedEncodingException, CannotSendEmailException {
+    public void sendMimeEmailWithFreemarker() throws UnsupportedEncodingException, CannotSendEmailException, URISyntaxException {
         InlinePicture inlinePicture = createGalaxyInlinePicture();
 
         final Email email = DefaultEmail.builder()
@@ -53,9 +54,9 @@ public class TestService {
         emailService.send(email, template, modelObject, inlinePicture);
     }
 
-    private InlinePicture createGalaxyInlinePicture() {
+    private InlinePicture createGalaxyInlinePicture() throws URISyntaxException {
         ClassLoader classLoader = getClass().getClassLoader();
-        File pictureFile = new File(classLoader.getResource("images" + File.separator + "galaxy.jpeg").getFile());
+        File pictureFile = new File(classLoader.getResource("images" + File.separator + "galaxy.jpeg").toURI());
         Preconditions.checkState(pictureFile.exists(), "There is not picture %s", pictureFile.getName());
 
         return DefaultInlinePicture.builder()