+++ spring-boot-email-template-engines/spring-boot-thymeleaf-email/src/test/java/it/ozimov/springboot/mail/templating/service/TemplatingTestUtils.java
@@ -20,6 +20,7 @@ import com.google.common.collect.ImmutableMap;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URISyntaxException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.util.Map;
@@ -36,9 +37,9 @@ public class TemplatingTestUtils {
             .put("activationLink", URL)
             .build();
 
-    public static String getExpectedBody() throws IOException {
+    public static String getExpectedBody() throws IOException, URISyntaxException {
         final File file = new File(ThymeleafTemplateServiceTest.class.getClassLoader()
-                .getResource("templates" + File.separator + TEMPLATE).getFile());
+                .getResource("templates" + File.separator + TEMPLATE).toURI());
         final String template = readFile(file);
         return template
                 .replace("<em th:text=\"${name}\">", "<em>" + NAME)